Finish Exercises 4.15-4.23, and Part 4 learnt about json web tokens, testing the routes, creating middleware

+++ part4/controllers/blogs.js
@@ -9,7 +9,7 @@ const Blog = require('../models/blog')
 // })
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1, id: 1 })
   response.json(blogs)
 })
 
@@ -25,24 +25,41 @@ blogsRouter.get('/:id', async (request, response) => {
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
 
+  const user = request.user
+  if(!user) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
   const blogToAdd = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
-    likes: body.likes ? body.likes : 0
+    likes: body.likes ? body.likes : 0,
+    user: user.id
   })
 
   if(body.title === undefined || body.url === undefined) {
     response.status(400).end()
   } else {
     const blogToSave = await blogToAdd.save()
+    user.blogs = user.blogs.concat(blogToSave._id)
+    await user.save()
     response.status(201).json(blogToSave)
   }
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const user = request.user
+  if(!user) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const blogToDelete = await Blog.findById(request.params.id)
+  if(blogToDelete.user.toString() === request.user.id) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    return response.status(401).json({ error: 'Not authorized to delete the blog' })
+  }
 })
 
 blogsRouter.put('/:id', async(request, response) => {
